+++ ORStools/proc/isochrones_point_proc.py
@@ -33,6 +33,8 @@ from qgis.core import (QgsWkbTypes,
                        QgsProcessingParameterString,
                        QgsProcessingParameterEnum,
                        QgsProcessingParameterPoint,
+                       QgsProcessingParameterNumber,
+                       QgsProcessingParameterDefinition
                        )
 
 from ORStools.common import isochrones_core, PROFILES, DIMENSIONS
@@ -51,6 +53,7 @@ class ORSIsochronesPointAlgo(ORSBaseProcessingAlgorithm):
         self.IN_RANGES = 'INPUT_RANGES'
         self.IN_KEY = 'INPUT_APIKEY'
         self.IN_DIFFERENCE = 'INPUT_DIFFERENCE'
+        self.IN_SMOOTHING = 'INPUT_SMOOTHING'
         self.PARAMETERS = [
             QgsProcessingParameterPoint(
                 name=self.IN_POINT,
@@ -67,8 +70,20 @@ class ORSIsochronesPointAlgo(ORSBaseProcessingAlgorithm):
                 name=self.IN_RANGES,
                 description=self.tr("Comma-separated ranges [min or m]"),
                 defaultValue="5, 10"
-            )
+            ),
+            QgsProcessingParameterNumber(
+                name=self.IN_SMOOTHING,
+                description=self.tr("Smoothing factor between 0 [detailed] and 100 [generalized]"),
+                defaultValue=None,
+                minValue=0,
+                maxValue=100,
+                optional=True
+            ),
         ]
+
+        # Set flag of smoothness parameters to advanced
+        smooth_param = next((i for i in self.PARAMETERS if i.name() == self.IN_SMOOTHING), None)
+        smooth_param.setFlags(smooth_param.flags() | QgsProcessingParameterDefinition.FlagAdvanced)
 
     # Save some important references
     # TODO bad style, refactor
@@ -88,6 +103,7 @@ class ORSIsochronesPointAlgo(ORSBaseProcessingAlgorithm):
         factor = 60 if dimension == 'time' else 1
         ranges_raw = parameters[self.IN_RANGES]
         ranges_proc = [x * factor for x in map(int, ranges_raw.split(','))]
+        smoothing = parameters[self.IN_SMOOTHING]
 
         options = self.parseOptions(parameters, context)
 
@@ -102,8 +118,11 @@ class ORSIsochronesPointAlgo(ORSBaseProcessingAlgorithm):
             "range": ranges_proc,
             "attributes": ['total_pop'],
             "id": None,
-            "options": options
+            "options": options,
         }
+
+        if smoothing or smoothing == 0:
+            params["smoothing"] = smoothing
 
         (sink, self.dest_id) = self.parameterAsSink(parameters, self.OUT, context,
                                                     self.isochrones.get_fields(),